refactor: fsd 구조에 알맞게 구조 변경
