+++ src/features/register-project/index.ts
@@ -1,0 +1,4 @@
+export * from './model/schema';
+export * from './model/usePostProjectLogo';
+export * from './model/usePostProjectRegistration';
+export { default as RegisterProjectForm } from './ui/RegisterProjectForm';
