+++ src/views/register/index.ts
@@ -1,4 +1,1 @@
-export * from './model/types';
-export * from './model/usePostProjectLogo';
-export * from './model/usePostProjectRegistration';
 export { default as RegisterPage } from './ui/RegisterPage';
