--- src/widgets/register/index.ts
@@ -1,3 +1,0 @@
-export * from './model/schema';
-export * from './style/textStyle';
-export { default as InputForm } from './ui/InputForm';
